# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Yet Another Node.js Blogg Application API",
      "description": "Yet Another Node.js Blogg Application API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "tags": [
      {
        "name": "Quote",
        "description": "API for Quote in the system"
      }
    ],
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/quote": {
        "post": {
          "tags": [
            "Quote"
          ],
          "description": "Get new quote in system",
          "parameters": [
            {
              "name": "user",
              "in": "body",
              "description": "User that we want to create",
              "schema": {
                "$ref": "#/definitions/Quote"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "New user is created",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "get": {
          "tags": [
            "Quote"
          ],
          "summary": "Get Quote in system",
          "parameters": [
            {
                "name": "fromTokenAddress",
                "in": "query",
                "description": "Quote check",
              "schema": {
                "$ref": "#/definitions/Quote"
              }
            },
              {
                "name": "toTokenAddress",
                "in": "query",
                "description": "Quote check",
              "schema": {
                "$ref": "#/definitions/Quote"
              }
            },
            {
                "name": "amount",
                "in": "query",
                "description": "Quote check",
              "schema": {
                "$ref": "#/definitions/Quote"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      }
    },
    "definitions": {
      "Quote": {
        "required": [
          "fromTokenAddress",
          "toTokenAddress",
          "amount"
        ],
        "properties": {
          "fromTokenAddress": {
            "type": "string",
            "uniqueItems": true
          },
          "toTokenAddress": {
            "type": "string",
            "uniqueItems": true
          },
          "amount": {
            "type": "string"
          }
        }
      },
      "Users": {
        "type": "array",
        "$ref": "#/definitions/Quote"
      }
    }
  }